refactor(database): update column type for penalty value in CreateOrdersTable migration

+++ src/database/migrations/1728551978379-CreateOrdersTable.ts
@@ -59,7 +59,7 @@ export class CreateOrdersTable1728551978379 implements MigrationInterface {
           },
           {
             name: 'penalty_value',
-            type: 'float',
+            type: 'decimal',
             precision: 10,
             scale: 2,
             isNullable: true,
